phase 8-1: lift calendar rangeFor to shared lib
