+++ frontend/src/routes/kids.$id.calendar.tsx
@@ -1,6 +1,5 @@
 import { useState, useMemo } from 'react';
 import { createFileRoute } from '@tanstack/react-router';
-import { startOfWeek, addDays, startOfMonth, endOfMonth, format } from 'date-fns';
 import type { View } from 'react-big-calendar';
 import { Skeleton } from '@/components/ui/skeleton';
 import { ErrorBanner } from '@/components/common/ErrorBanner';
@@ -8,28 +7,10 @@ import { useKid, useKidCalendar } from '@/lib/queries';
 import { KidTabs } from '@/components/layout/KidTabs';
 import { CalendarView } from '@/components/calendar/CalendarView';
 import { CalendarEventPopover } from '@/components/calendar/CalendarEventPopover';
+import { rangeFor } from '@/lib/calendarRange';
 import type { CalendarEvent } from '@/lib/types';
 
 export const Route = createFileRoute('/kids/$id/calendar')({ component: KidCalendarPage });
-
-const BUFFER_DAYS = 3;
-
-function rangeFor(view: View, cursor: Date): { from: string; to: string } {
-  if (view === 'month') {
-    const monthStart = startOfMonth(cursor);
-    const monthEnd = endOfMonth(cursor);
-    const weekStart = startOfWeek(monthStart, { weekStartsOn: 0 });
-    return {
-      from: format(addDays(weekStart, -BUFFER_DAYS), 'yyyy-MM-dd'),
-      to: format(addDays(monthEnd, 7 + BUFFER_DAYS), 'yyyy-MM-dd'),
-    };
-  }
-  const weekStart = startOfWeek(cursor, { weekStartsOn: 0 });
-  return {
-    from: format(addDays(weekStart, -BUFFER_DAYS), 'yyyy-MM-dd'),
-    to: format(addDays(weekStart, 7 + BUFFER_DAYS), 'yyyy-MM-dd'),
-  };
-}
 
 function KidCalendarPage() {
   const { id } = Route.useParams();
